+++ ct/umami.sh
@@ -1,9 +1,9 @@
 #!/usr/bin/env bash
 source <(curl -fsSL https://raw.githubusercontent.com/community-scripts/ProxmoxVE/main/misc/build.func)
-# Copyright (c) 2021-2025 tteck
+# Copyright (c) 2021-2026 tteck
 # Author: tteck (tteckster)
 # License: MIT | https://github.com/community-scripts/ProxmoxVE/raw/main/LICENSE
-# Source: https://umami.is/
+# Source: https://umami.is/ | Github: https://github.com/umami-software/umami
 
 APP="Umami"
 var_tags="${var_tags:-analytics}"
@@ -11,7 +11,7 @@ var_cpu="${var_cpu:-2}"
 var_ram="${var_ram:-2048}"
 var_disk="${var_disk:-12}"
 var_os="${var_os:-debian}"
-var_version="${var_version:-12}"
+var_version="${var_version:-13}"
 var_unprivileged="${var_unprivileged:-1}"
 
 header_info "$APP"
@@ -20,39 +20,40 @@ color
 catch_errors
 
 function update_script() {
-    header_info
-    check_container_storage
-    check_container_resources
-    if [[ ! -d /opt/umami ]]; then
-        msg_error "No ${APP} Installation Found!"
-        exit
-    fi
+  header_info
+  check_container_storage
+  check_container_resources
+  if [[ ! -d /opt/umami ]]; then
+    msg_error "No ${APP} Installation Found!"
+    exit
+  fi
 
-    msg_info "Stopping ${APP}"
+  if check_for_gh_release "umami" "umami-software/umami"; then
+    msg_info "Stopping Service"
     systemctl stop umami
-    msg_ok "Stopped $APP"
+    msg_ok "Stopped Service"
 
     fetch_and_deploy_gh_release "umami" "umami-software/umami" "tarball"
 
-    msg_info "Updating ${APP}"
+    msg_info "Updating Umami"
     cd /opt/umami
-    $STD yarn install
-    $STD yarn run build
-    msg_ok "Updated ${APP}"
+    $STD pnpm install
+    $STD pnpm run build
+    msg_ok "Updated Umami"
 
-    msg_info "Starting ${APP}"
+    msg_info "Starting Service"
     systemctl start umami
-    msg_ok "Started ${APP}"
-
-    msg_ok "Updated Successfully"
-    exit
+    msg_ok "Started Service"
+    msg_ok "Updated successfully!"
+  fi
+  exit
 }
 
 start
 build_container
 description
 
-msg_ok "Completed Successfully!\n"
+msg_ok "Completed successfully!\n"
 echo -e "${CREATING}${GN}${APP} setup has been successfully initialized!${CL}"
 echo -e "${INFO}${YW} Access it using the following URL:${CL}"
 echo -e "${TAB}${GATEWAY}${BGN}http://${IP}:3000${CL}"
